+++ projects/server/src/controllers/products/units/index.js
@@ -1,4 +1,3 @@
-const {Op} = require("sequelize")
 const {middlewareErrorHandling} = require("../../../middleware/index.js");
 const { Product_Unit, Product_Detail } = require("../../../model/relation.js")
 const {productUnitValidationSchema, deleteProductUnitValidationSchema } = require("./validation.js")
@@ -190,49 +189,9 @@ const deleteProductUnits = async( req, res, next ) => {
   }
 }
 
-const makeConvertionUnit = async( req, res, next ) => {
-  try{
-    
-    const isProductDefaultAvailable = await Product_Detail.findOne({
-      where : {
-        productId : req.body.productId,
-        isDefault : 1,
-        quantity : {
-          [Op.gt] : 0
-        }
-      }
-    })
-    
-    if(!isProductDefaultAvailable) throw ({
-      status : middlewareErrorHandling.NOT_FOUND_STATUS, 
-      message : middlewareErrorHandling.CANNOT_MAKING_CONVERTION 
-    })
-
-    await Product_Detail.update(
-      { quantity : isProductDefaultAvailable.dataValues?.quantity - req.body.times },
-      { where : { stockId : isProductDefaultAvailable?.dataValues?.stockId } }
-    )
-
-    const unitConvertion = await Product_Detail.update(
-      { quantity : (isProductDefaultAvailable.dataValues?.convertion * req.body.times ) },
-      { where : { productId : req.body.productId, isDefault : 0 } }
-    )
-
-    res.status(200).json({ 
-      type : "success",
-      message : "Unit Produk berhasil dikonversi",
-      unitConvertion
-    })
-
-  } catch (error){
-    next(error)
-  }
-}
-
 module.exports = {
     productUnits,
     setProductUnits,
     updateProductUnits,
-    deleteProductUnits,
-    makeConvertionUnit
+    deleteProductUnits
 }
